requirements.txt
altgraph==0.16.1
amqp==2.4.2
anyjson==0.3.3
asn1crypto==0.24.0
backports.csv==1.0.7
billiard==3.5.0.5
celery==4.2.1
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
cipher==0.1
Click==7.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.6.1
defusedxml==0.6.0
diff-match-patch==20181111
dingtalk==0.0.5
Django==2.2.1
django-celery-results==1.0.4
django-crispy-forms==1.7.2
django-filter==2.1.0
django-formtools==2.1
django-guardian==1.5.1
django-import-export==1.2.0
django-multiselectfield==0.1.8
django-pure-pagination==0.3.0
django-ranged-response==0.2.0
django-reversion==3.0.3
django-simple-captcha==0.5.11
djangorestframework==3.9.4
dnspython==1.16.0
et-xmlfile==1.0.1
eventlet==0.24.1
Flask==1.0.2
future==0.17.1
greenlet==0.4.15
httplib2==0.9.2
idna==2.8
itsdangerous==1.1.0
itypes==1.1.0
jdcal==1.4.1
Jinja2==2.10.1
kombu==4.4.0
macholib==1.11
Markdown==3.1
MarkupSafe==1.1.1
monotonic==1.5
mysqlclient==1.4.2.post1
odfpy==1.4.0
openpyxl==2.6.2
pefile==2018.8.8
Pillow==6.0.0
pkcs7==0.1.2
psycopg2==2.8.2
pycparser==2.19
pycrypto==2.6.1
pycryptodome==3.7.3
Pygments==2.4.0
PyInstaller==3.4
PyMySQL==0.9.3
PyQt5==5.12
PyQt5-sip==4.19.14
pytz==2019.1
pywin32-ctypes==0.2.0
PyYAML==5.1
redis==3.2.0
requests==2.21.0
requirements==0.1
reversion==0.2
six==1.12.0
sqlparse==0.3.0
tablib==0.13.0
toml==0.10.0
uritemplate==3.0.0
urllib3==1.24.3
vine==1.3.0
Werkzeug==0.15.2
xlrd==1.2.0
xlwt==1.3.0
